Configure with OCI8 instantclient link hardcoded.
Previously, this was relying on ORACLE_HOME being set (the
oracle-instantclient packages create a profile.d file to do this).  This
was poor because someone trying to build it may not (yet) have executed
the profile.d script and won't have the ORACLE_HOME environment variable
set.

Because this reliance on ORACLE_HOME is gone, I'm removing it from the
Dockerfile as well.